+++ src/main/java/ru/kolaer/permit/dao/PermitStatusHistoryPageDao.java
@@ -8,9 +8,9 @@ import java.util.List;
  * Created by Danilov on 27.05.2017.
  */
 public interface PermitStatusHistoryPageDao extends BasePageDao<PermitStatusHistoryEntity> {
-    PermitStatusHistoryEntity findLastStatusByIdPermit(Integer id);
+    PermitStatusHistoryEntity findLastStatusByIdPermit(Long id);
 
-    List<PermitStatusHistoryEntity> findLastStatusByIdPermitRange(List<Integer> id);
+    List<PermitStatusHistoryEntity> findLastStatusByIdPermitRange(List<Long> id);
 
-    List<PermitStatusHistoryEntity> findAllById(Integer id);
+    List<PermitStatusHistoryEntity> findAllById(Long id);
 }
